requirements.txt
agent-os-kernel==3.4.0
agentmesh-platform==3.4.0
agent-governance-toolkit==3.4.0